add mobileview option (replaces count and enableNavigator).
Understandably, the code I added to detect mobile is pretty damn ugly in `charts.js`, but the options passed to `addNewCharts` method is not made available to the `charts` object. Making the code intuitive is tricky due to the design of the `charts` abstraction: each chart is not seen as an individual object, but a single `charts` controls all charts. This is why the element id needs to passed for every single function call in `charts`. Attributes of each chart is then attached to DOM elements via jQuery's `data()` method, which currently only assumes a single attribute `options`.

Yes, this is not good OOP design, but god knows how long refactoring will take... so this will do (: